AI_django/requirements.txt
affine==2.3.1
amqp==5.1.1
asgiref==3.5.1
attrs==21.4.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.24.14
botocore==1.27.14
cachetools==5.0.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cligj==0.7.2
cycler==0.11.0
Django==3.2.13
django-celery-beat==2.2.1
django-cors-headers==3.13.0
django-storages==1.12.3
django-timezone-field==4.2.3
djangorestframework==3.13.1
earthengine-api==0.1.309
easydict==1.9
Fiona==1.8.21
fonttools==4.33.3
future==0.18.2
geojson==2.5.0
geomet==0.3.0
google-api-core==2.7.3
google-api-python-client==1.12.11
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.1
gunicorn==20.1.0
html2text==2020.1.16
httplib2==0.20.4
httplib2shim==0.0.3
idna==3.3
jdatetime==4.1.0
jmespath==1.0.1
joblib==1.1.0
kiwisolver==1.4.3
kombu==5.2.4
munch==2.5.0
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
rasterio==1.2.10
requests==2.27.1
rsa==4.8
s3transfer==0.6.0
scikit-learn==1.1.1
scipy==1.8.1
sentinelsat==1.1.1
Shapely==1.8.2
six==1.16.0
snuggs==1.4.7
sqlparse==0.4.2
threadpoolctl==3.1.0
tifffile==2022.5.4
torch
torchaudio
torchvision
tqdm==4.64.0
typing_extensions==4.2.0
uritemplate==3.0.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
